+++ src/apps/explorer/ExplorerApp.tsx
@@ -7,8 +7,10 @@ import useNetworkCheck from 'hooks/useNetworkCheck'
 import Console from 'Console'
 import { rootReducer, INITIAL_STATE } from 'apps/explorer/state'
 
+import styled from 'styled-components'
 import { GenericLayout } from 'components/layout'
 import { Header } from './layout/Header'
+import { media } from 'theme/styles/media'
 
 import { NetworkUpdater } from 'state/network'
 import { initAnalytics } from 'api/analytics'
@@ -95,6 +97,20 @@ const RedirectMainnet = (): JSX.Element => {
   return <Redirect push={false} to={newPath} />
 }
 
+const Wrapper = styled.div`
+  max-width: 140rem;
+  margin: 0 auto;
+
+  ${media.mediumDown} {
+    max-width: 94rem;
+    flex-flow: column wrap;
+  }
+
+  ${media.mobile} {
+    max-width: 100%;
+  }
+`
+
 /**
  * Render Explorer App
  */
@@ -103,7 +119,7 @@ export const ExplorerApp: React.FC = () => {
   useNetworkCheck()
 
   return (
-    <>
+    <Wrapper>
       <Router basename={process.env.BASE_URL}>
         <StateUpdaters />
         <Switch>
@@ -112,7 +128,7 @@ export const ExplorerApp: React.FC = () => {
         </Switch>
       </Router>
       {process.env.NODE_ENV === 'development' && <Console />}
-    </>
+    </Wrapper>
   )
 }
