+++ src/apps/explorer/pages/Home/index.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import styled from 'styled-components'
-import { media } from 'theme/styles/media'
 import { Search } from 'apps/explorer/components/common/Search'
 
 const Wrapper = styled.div`
@@ -11,15 +10,6 @@ const Wrapper = styled.div`
   padding: 1.6rem;
   margin: 0 auto;
   width: 100%;
-  max-width: 140rem;
-
-  ${media.mediumDown} {
-    max-width: 94rem;
-  }
-
-  ${media.mobile} {
-    max-width: 100%;
-  }
 
   > h1 {
     display: flex;
